修复avcodec inner同步接口输入buffer size类型范围超过实际使用范围的问题

+++ interfaces/inner_api/native/avcodec_audio_codec.h
@@ -81,7 +81,7 @@ public:
         (void)bufferStatus;
     }
 
-    virtual int32_t QueryInputBuffer(uint32_t *index, size_t bufferSize, int64_t timeoutUs)
+    virtual int32_t QueryInputBuffer(uint32_t *index, int32_t bufferSize, int64_t timeoutUs)
     {
         (void)index;
         (void)bufferSize;
